+++ app-main/src/main/java/net/causw/app/main/domain/notification/notification/service/implementation/UserPushTokenWriter.java
@@ -1,4 +1,4 @@
-package net.causw.app.main.domain.user.account.service.implementation;
+package net.causw.app.main.domain.notification.notification.service.implementation;
 
 import java.util.HashSet;
 import java.util.Set;
