+++ app-main/src/main/java/net/causw/app/main/shared/infra/firebase/FcmUtils.java
@@ -3,6 +3,7 @@ package net.causw.app.main.shared.infra.firebase;
 import java.util.HashSet;
 import java.util.Set;
 
+import net.causw.app.main.domain.notification.notification.service.implementation.UserPushTokenWriter;
 import org.springframework.stereotype.Component;
 
 import net.causw.app.main.domain.user.account.entity.user.User;
@@ -13,7 +14,7 @@ import net.causw.global.constant.StaticValue;
 import lombok.RequiredArgsConstructor;
 
 /**
- * @deprecated V2 이후 신규 코드는 {@link net.causw.app.main.domain.user.account.service.implementation.UserPushTokenWriter}를 사용하세요. V1 및 배치 호환을 위해 유지됩니다.
+ * @deprecated V2 이후 신규 코드는 {@link UserPushTokenWriter}를 사용하세요. V1 및 배치 호환을 위해 유지됩니다.
  */
 @Deprecated
 @RequiredArgsConstructor
